requirements.txt
Flask
flask-cors
fpdf
python-docx
